backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
dnspython==2.4.2
